requirements.txt
pytest
libcamera
picamera2
PyYAML==6.0
pillow
gpiozero
functools
numpy
black
paho-mqtt
pillow
python-dateutil
pytz
pdocs
pybase64